==> RAG_pipeline_pdf/7_ragas.ipynb <==
from dotenv import load_dotenv

import os

load_dotenv()
ASU_key = os.environ.get("ASU_key") 
endpoint_url = os.environ.git("endpoint_url")

import requests
endpoint_url = 'https://api-dev-poc.aiml.asu.edu/eval'
headers = {
    "Authorization": f"Bearer {ASU_key}",
    "Content-Type": "application/json"
}
### Bluescore
payload_blue = {
    "model_provider": "openai",
    "model_name": "gpt-4o",
    "metric": "bleuscore",
    "parameters": {
        "user_input": "summarise given text\nThe company reported an 8% rise in Q3 2024, driven by strong performance in the Asian market. Sales in this region have significantly contributed to the overall growth. Analysts attribute this success to strategic marketing and product localization. The positive trend in the Asian market is expected to continue into the next quarter.",
        "response": "The company experienced an 8% increase in Q3 2024, largely due to effective marketing strategies and product adaptation, with expectations of continued growth in the coming quarter.",
        "reference": "The company reported an 8% growth in Q3 2024, primarily driven by strong sales in the Asian market, attributed to strategic marketing and localized products, with continued growth anticipated in the next quarter."
    }
}

try:
    response = requests.post(endpoint_url, headers=headers, json=payload_blue)
    response.raise_for_status()
    result = response.json()
    print("result:", result)
except requests.exceptions.RequestException as e:
    print(f"API request error: {e}")
except Exception as e:
    print(f"Unexpected error: {e}")

# LLM Evaluation payloads

### Context Precision With Reference
{
    "model_provider": "openai",
    "model_name": "gpt-4o",
    "metric": "context_precision_with_reference",
    "parameters": {
        "user_input": "Where is the Eiffel Tower located?",
        "reference": "The Eiffel Tower is located in Paris.",
        "retrieved_contexts": [
            "The Eiffel Tower is located in Paris."
        ]
    }
}

### Context Precision with Response
{
    "model_provider": "openai",
    "model_name": "gpt-4o",
    "metric": "context_precision_with_response",
    "parameters": {
        "user_input": "Where is the Eiffel Tower located?",
        "response": "The Eiffel Tower is located in Paris.",
        "retrieved_contexts": [
            "The Eiffel Tower is located in Paris."
        ]
    }
}

### Non llm context precision
{
    "model_provider": "openai",
    "model_name": "gpt-4o",
    "metric": "nonllm_context_precision",
    "parameters": {
        "retrieved_contexts": [
            "The Eiffel Tower is located in Paris."
        ],
        "reference_contexts": [
            "Paris is the capital of France.",
            "The Eiffel Tower is one of the most famous landmarks in Paris."
        ]
    }
}

### Faithfulness
{
    "model_provider": "openai",
    "model_name": "gpt-4o",
    "metric": "faithfulness",
    "parameters": {
        "user_input": "When was the first super bowl?",
        "response": "The first superbowl was held on Jan 15, 1967",
        "retrieved_contexts": [
            "The First AFL\u2013NFL World Championship Game was an American football game played on January 15, 1967, at the Los Angeles Memorial Coliseum in Los Angeles."
        ]
    }
}
```
### bleuscore
{
    "model_provider": "openai",
    "model_name": "gpt-4o",
    "metric": "bleuscore",
    "parameters": {
        "user_input": "summarise given text\nThe company reported an 8% rise in Q3 2024, driven by strong performance in the Asian market. Sales in this region have significantly contributed to the overall growth. Analysts attribute this success to strategic marketing and product localization. The positive trend in the Asian market is expected to continue into the next quarter.",
        "response": "The company experienced an 8% increase in Q3 2024, largely due to effective marketing strategies and product adaptation, with expectations of continued growth in the coming quarter.",
        "reference": "The company reported an 8% growth in Q3 2024, primarily driven by strong sales in the Asian market, attributed to strategic marketing and localized products, with continued growth anticipated in the next quarter."
    }
}
### Factual correctness
```json
{
    "model_provider": "openai",
    "model_name": "gpt-4o",
    "metric": "factual_correctness",
    "parameters": {
        "response": "The Eiffel Tower is located in Paris.",
        "reference": "The Eiffel Tower is located in Paris. I has a height of 1000ft."
    }
}
```
### Semantic similarity
```json
{
    "model_provider": "openai",
    "model_name": "gpt-4o",
    "metric": "semantic_similarity",
    "parameters": {
        "response": "The Eiffel Tower is located in Paris.",
        "reference": "The Eiffel Tower is located in Paris. It has a height of 1000ft."
    }
}
```
### Non llm string similarity
```json
{
    "model_provider": "openai",
    "model_name": "gpt-4o",
    "metric": "non_llm_string_similarity",
    "parameters": {
        "response": "The Eiffel Tower is located in Paris.",
        "reference": "The Eiffel Tower is located in Paris. It has a height of 1000ft."
    }
}
```
### Rouge score
```json
{
    "model_provider": "openai",
    "model_name": "gpt-4o",
    "metric": "rouge_score",
    "parameters": {
        "response": "The Eiffel Tower is located in Paris.",
        "reference": "The Eiffel Tower is located in Paris. It has a height of 1000ft."
    }
}
