==> RAG_pipeline_pdf/8_webpage_scrape.ipynb <==
!pip install langdetect

import pandas as pd

df = pd.read_csv('/Users/sarahherberger/Documents/ASU/xavier_project/webpage_analysis.csv')

df.head()

import pandas as pd
import re

def clean_text(text):
  """Keep only alphanumeric characters and basic punctuation."""
  return re.sub(r'[^a-zA-Z0-9.,!?\s]', '', str(text))

# Apply the function to the DataFrame column
df['cleaned_text'] = df['page_text'].apply(clean_text)
df['title'] = df['title'].apply(clean_text)
df['topic'] = df['topic'].apply(clean_text)

from langdetect import detect, DetectorFactory
DetectorFactory.seed = 0

def lang_detect(text):
  text = re.sub(r'[^a-zA-Z0-9.,!?\s]', '', str(text))
  language_detected = detect(text)
  return language_detected

df['langauage'] = df['cleaned_text'].apply(lang_detect)

df.columns

print(df.shape)
df = df[df['word_count'] > 50]
print(df.shape)
df = df[df['langauage'] == 'en']
print(df.shape)
df_clean = df.drop_duplicates(subset=['title', 'topic', 'word_count', 'char_count', 'cleaned_text', 'langauage'])
print(df_clean.shape)

df_clean[['title', 'url', 'cleaned_text']].head()

df_clean[['title', 'url', 'cleaned_text']].to_csv('cleaned_ASU_webpage.csv')

df
